Ok working on new feats here.
* Added a client list whichs keeps track of dues.
* Purchase dialog has new "paying field", so to specify how much the
costumer is paying. It's automatically set to the total, but if changed
a new dialog will pop-up to ask who is gonna owe the difference.
* Did a few changes on Auto Complete Text Field, it has a sort of
temporary "autoProc" boolean which, if true, will make the field call
it's onAction when context menu is hidden. Product fields are auto-proc,
whereas others like client name field, which could be a brand new
client, are not.

TODOs:
* SAVE THE CLIENT LIST!
* get paid
* Fix goddamn autocompletefield thing