History and client overhaul.
* Changed savefile by a lot, but code to adapt old saves was added.
* Client debt is no longer an arbitrary number, and is calculated
directly from all the transactions with his name.
* Purchase class removed. It's now a subclass of the new Transaction
class.
* Payment is another subclass from transaction. And a legacy debt to
save the debt that was lost on the overhaul.
* Only purchase class contains a product list. Others just have creation
date, balance, client name and whether it's cancelled or not.
* Instead of "undone", transactions now can only be "nullified". They
stay in history but do not count towards calculations.
* Save file now saves the version of the program it was made by, so to
better manage savefile updates on later versions.

Note: due to the huge changes this version is very prone to bugs. It requires heavy testing...